+++ examples/special-types/src/main.rs
@@ -1,5 +1,3 @@
-use std::path::PathBuf;
-
 use rocket::http::Status;
 use rocket::{get, post, serde::json::Json};
 use rocket_okapi::okapi::schemars;
@@ -7,6 +5,7 @@ use rocket_okapi::settings::UrlObject;
 use rocket_okapi::{openapi, openapi_get_routes, rapidoc::*, swagger_ui::*};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
 
 /// # Get data
 #[openapi(tag = "Users")]